// src/features/catalog/domain/models/Customer.ts
export type Customer = {
  id: number
  identification_document_code: string
  identification: string
  dv: string
  legal_organization_code: string
  tribute_code: string
  responsibilities: string[]
  company: string
  trade_name: string
  names: string
  address: string
  email: string
  phone: string
  country_code: string
  municipality_code: string
}